requirements.txt
streamlit
pymupdf
google-genai